refactor: jpa audit 정리
