--- src/main/java/org/prography/kagongsillok/common/auditing/AuditingTimeEntity.java
@@ -1,34 +1,0 @@
-package org.prography.kagongsillok.common.auditing;
-
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import java.time.ZonedDateTime;
-import javax.persistence.EntityListeners;
-import javax.persistence.MappedSuperclass;
-import javax.persistence.PrePersist;
-import javax.persistence.PreUpdate;
-import lombok.Data;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@Data
-@MappedSuperclass
-@EntityListeners(value = AuditingEntityListener.class)
-public class AuditingTimeEntity {
-
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss.SSS", timezone = "Asia/Seoul")
-    private ZonedDateTime createdAt;
-
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss.SSS", timezone = "Asia/Seoul")
-    private ZonedDateTime updatedAt;
-
-    @PrePersist
-    public void prePersist() {
-        this.createdAt = ZonedDateTime.now();
-        this.updatedAt = ZonedDateTime.now();
-    }
-
-    @PreUpdate
-    public void preUpdate() {
-        this.updatedAt = ZonedDateTime.now();
-    }
-}
