+++ src/main/java/org/prography/kagongsillok/common/entity/AbstractRootEntity.java
@@ -1,0 +1,28 @@
+package org.prography.kagongsillok.common.entity;
+
+
+import java.time.ZonedDateTime;
+import javax.persistence.EntityListeners;
+import javax.persistence.MappedSuperclass;
+import lombok.Getter;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+@Getter
+@MappedSuperclass
+@EntityListeners(value = AuditingEntityListener.class)
+public class AbstractRootEntity {
+
+    @CreationTimestamp
+    private ZonedDateTime createdAt;
+
+    @UpdateTimestamp
+    private ZonedDateTime updatedAt;
+
+    private Boolean isDeleted = Boolean.FALSE;
+
+    public void delete() {
+        this.isDeleted = Boolean.TRUE;
+    }
+}
