+++ src/main/java/org/prography/kagongsillok/place/application/exception/NotFoundPlaceException.java
@@ -2,7 +2,7 @@ package org.prography.kagongsillok.place.application.exception;
 
 import org.prography.kagongsillok.common.exception.NotFoundException;
 
-public class NotFoundPlaceException extends NotFoundException {
+public final class NotFoundPlaceException extends NotFoundException {
 
     public NotFoundPlaceException(final Long placeId) {
         super(String.format("장소가 존재하지 않습니다. placeId = %d", placeId));
