+++ src/main/java/org/prography/kagongsillok/review/ui/dto/ReviewUpdateRequest.java
@@ -50,5 +50,4 @@ public class ReviewUpdateRequest {
                 .reviewTagIds(reviewTagIds)
                 .build();
     }
-
 }
